Cover script is working!
Cover script is assigned as a component to the player. It allows the player to take cover against a nearby object, and locks their movement to that axis. When the player is near the edge of a cover object, it sets the boolean variable atCornerCover to true, which will later allow PlayerWalking script to stop the player from moving further.