Cover: Made it so you can't move past corners of cover
Before, when the player was in cover against a wall and attempted to move past the edge, they would orbit around the corner and stick out from around the corner. I added a boolean variable in CoverScript.cs to determine which corner the player is at, if any, and some additional logic in PlayerWalking.cs to stop the player from moving around corners when they're in cover.